swagger接口

+++ src/main/java/global/com/dao/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends CrudRepository<User, Long> {
-    public User findUserById(Long ById) ;
+    public User findUserById(Integer ById) ;
 
 
 }
