+++ src/main/java/global/com/service/EcacheService.java
@@ -4,5 +4,5 @@ import global.com.dao.User;
 
 public interface EcacheService {
     public User saveLocalCache(User user);
-    public User getLocalCache(Long  id);
+    public User getLocalCache(Integer  id);
 }
